employ useConnect pattern
