+++ frontend/src/views/DifferentialExpression/components/Main/components/DeResults/components/DifferentialExpressionResults/index.tsx
@@ -1,4 +1,4 @@
-import { useState, useMemo, SetStateAction, Dispatch } from "react";
+import { useMemo } from "react";
 import {
   CellGroupTitle,
   CellGroupTitleWrapper,
@@ -15,15 +15,10 @@ import {
   StyledTooltipText,
 } from "./style";
 import cxgIcon from "./images/cxg.svg";
-import {
-  QueryGroup,
-  QueryGroups,
-} from "src/views/DifferentialExpression/common/store/reducer";
 import { Pagination } from "@mui/material";
 import Table from "src/views/CellGuide/components/CellGuideCard/components/common/Table";
 import { ButtonIcon, Tooltip } from "@czi-sds/components";
 
-import { generateAndCopyShareUrl } from "./utils";
 import { DifferentialExpressionRow } from "../../types";
 import { MAX_NUM_TOP_GENES_TO_PORT_TO_GE, ROWS_PER_PAGE } from "./constants";
 import { CellCountTitle } from "../../../../style";
@@ -41,22 +36,11 @@ import {
   DIFFERENTIAL_EXPRESSION_RESULTS_TABLE,
   DIFFERENTIAL_EXPRESSION_SORT_DIRECTION,
 } from "src/views/DifferentialExpression/common/constants";
-import useProcessedQueryGroupFilterDimensions from "../../../common/query_group_filter_dimensions";
 import { track } from "src/common/analytics";
 import { EVENTS } from "src/common/analytics/events";
+import { useConnect } from "./connect";
+import { Props } from "./types";
 
-interface DifferentialExpressionResultsProps {
-  queryGroups: QueryGroups;
-  queryGroupsWithNames: QueryGroups;
-  organismId: string;
-  sortedAndFilteredResults: DifferentialExpressionRow[];
-  nCellsOverlap: number;
-  setSearchQuery: Dispatch<SetStateAction<string>>;
-  setLfcFilter: Dispatch<SetStateAction<string>>;
-  setEffectSizeFilter: Dispatch<SetStateAction<string>>;
-  sortDirection: "asc" | "desc";
-  setSortDirection: Dispatch<SetStateAction<"asc" | "desc">>;
-}
 const DifferentialExpressionResults = ({
   queryGroups,
   queryGroupsWithNames,
@@ -68,65 +52,25 @@ const DifferentialExpressionResults = ({
   setEffectSizeFilter,
   sortDirection,
   setSortDirection,
-}: DifferentialExpressionResultsProps) => {
-  const [page, setPage] = useState(1);
-
-  const { n_cells: nCellsGroup1 } = useProcessedQueryGroupFilterDimensions(
-    queryGroups.queryGroup1
-  );
-  const { n_cells: nCellsGroup2 } = useProcessedQueryGroupFilterDimensions(
-    queryGroups.queryGroup2
-  );
-  const handlePageChange = (
-    _event: React.ChangeEvent<unknown>,
-    page: number
-  ) => {
-    setPage(page);
-  };
-
-  const [openInGeHref1, openInGeHref2] = useMemo(() => {
-    const generateUrl = (
-      queryGroup: QueryGroup,
-      queryGroupWithNames: QueryGroup,
-      sliceFromBeginning: boolean
-    ) => {
-      return organismId
-        ? generateAndCopyShareUrl({
-            queryGroup: queryGroup,
-            organism: organismId,
-            genes: sliceFromBeginning
-              ? sortedAndFilteredResults
-                  .slice(0, MAX_NUM_TOP_GENES_TO_PORT_TO_GE)
-                  .map((row) => row.name)
-              : sortedAndFilteredResults
-                  .slice(-MAX_NUM_TOP_GENES_TO_PORT_TO_GE)
-                  .reverse()
-                  .map((row) => row.name),
-            cellTypes: queryGroupWithNames.cellTypes,
-          })
-        : "";
-    };
-    return [
-      generateUrl(
-        queryGroups.queryGroup1,
-        queryGroupsWithNames.queryGroup1,
-        sortDirection === "desc"
-      ),
-      generateUrl(
-        queryGroups.queryGroup2,
-        queryGroupsWithNames.queryGroup2,
-        sortDirection === "asc"
-      ),
-    ];
-  }, [
+}: Props) => {
+  const {
+    page,
+    setPage,
+    nCellsGroup1,
+    nCellsGroup2,
+    openInGeHref1,
+    openInGeHref2,
+    pageCount,
+    handlePageChange,
+    overlapPercent,
+  } = useConnect({
     queryGroups,
     queryGroupsWithNames,
     organismId,
-    sortDirection,
     sortedAndFilteredResults,
-  ]);
-
-  const pageCount = Math.ceil(sortedAndFilteredResults.length / ROWS_PER_PAGE);
+    nCellsOverlap,
+    sortDirection,
+  });
 
   const columnIdToName: Record<
     keyof Omit<DifferentialExpressionRow, "adjustedPValue">,
@@ -203,11 +147,9 @@ const DifferentialExpressionResults = ({
     setLfcFilter,
     setEffectSizeFilter,
     setSortDirection,
+    setPage,
   ]);
-  const overlapPercent = (
-    (nCellsOverlap / Math.max(nCellsGroup1, nCellsGroup2)) *
-    100
-  ).toFixed(2);
+
   return (
     <>
       <CellGroupWrapper data-testid={DIFFERENTIAL_EXPRESSION_CELL_GROUP_1_INFO}>
