+++ frontend/src/views/DifferentialExpression/components/Main/components/DeResults/types.ts
@@ -4,3 +4,7 @@ export interface DifferentialExpressionRow {
   effectSize: string;
   adjustedPValue: string;
 }
+
+export interface Props {
+  setIsLoading: (isLoading: boolean) => void;
+}
