+++ frontend/src/views/DifferentialExpression/components/Main/components/Filters/components/CopyButton/types.ts
@@ -1,0 +1,6 @@
+import { QueryGroup } from "src/views/DifferentialExpression/common/store/reducer";
+
+export interface Props {
+  queryGroupKey: keyof QueryGroup;
+  testId?: string;
+}
