+++ frontend/src/views/DifferentialExpression/components/Main/components/Filters/components/FilterDropdown/types.ts
@@ -1,0 +1,9 @@
+import { FilterOption } from "../../types";
+
+export interface Props {
+  label: string;
+  options: FilterOption[];
+  allAvailableOptions: FilterOption[];
+  selectedOptionIds: string[];
+  handleChange: (options: FilterOption[]) => void;
+}
