+++ frontend/src/views/DifferentialExpression/components/Main/components/Filters/types.ts
@@ -1,5 +1,12 @@
+import { QueryGroup } from "src/views/DifferentialExpression/common/store/reducer";
+
 export interface FilterOption {
   name: string;
   id: string;
   unavailable?: boolean;
+}
+
+export interface Props {
+  queryGroup: QueryGroup;
+  isQueryGroup1: boolean;
 }
