+++ frontend/src/views/DifferentialExpression/components/Main/index.tsx
@@ -1,8 +1,3 @@
-import { useContext, useEffect, useState } from "react";
-import {
-  DispatchContext,
-  StateContext,
-} from "src/views/DifferentialExpression/common/store";
 import {
   Wrapper,
   StepHeader,
@@ -19,10 +14,8 @@ import {
 } from "./style";
 import QueryGroupFilters from "./components/Filters";
 import Organism from "./components/Organism";
-import { submitQueryGroups } from "src/views/DifferentialExpression/common/store/actions";
 import DeResults from "./components/DeResults";
 import Loader from "./components/Loader";
-import useProcessedQueryGroupFilterDimensions from "./components/common/query_group_filter_dimensions";
 import Method from "./components/Method";
 import {
   DIFFERENTIAL_EXPRESSION_CELL_GROUP_1_FILTER,
@@ -34,44 +27,21 @@ import {
 } from "../../common/constants";
 import { track } from "src/common/analytics";
 import { EVENTS } from "src/common/analytics/events";
-import { craftPayloadWithQueryGroups } from "./utils";
+import { useConnect } from "./connect";
 
 export default function DifferentialExpression(): JSX.Element {
-  const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [isLoadingGetDeQuery, setIsLoadingGetDeQuery] =
-    useState<boolean>(false);
-
-  useEffect(() => {
-    setIsLoadingGetDeQuery(isLoadingGetDeQuery);
-  }, [isLoadingGetDeQuery]);
-  const dispatch = useContext(DispatchContext);
-  const { queryGroups } = useContext(StateContext);
-  const { queryGroup1, queryGroup2 } = queryGroups;
-
-  // check if any values in queryGroup1 are not empty
-  const isQueryGroup1NotEmpty = Object.values(queryGroup1).some(
-    (value) => value.length > 0
-  );
-  const isQueryGroup2NotEmpty = Object.values(queryGroup2).some(
-    (value) => value.length > 0
-  );
-  const canRunDifferentialExpression =
-    !isLoading && isQueryGroup1NotEmpty && isQueryGroup2NotEmpty;
-
-  const handleRunDifferentialExpression = () => {
-    if (!dispatch) return;
-    dispatch(submitQueryGroups());
-
-    track(
-      EVENTS.DE_FIND_GENES_CLICKED,
-      craftPayloadWithQueryGroups(queryGroups)
-    );
-  };
-
-  const { n_cells: nCellsGroup1, isLoading: isLoadingGroup1 } =
-    useProcessedQueryGroupFilterDimensions(queryGroup1);
-  const { n_cells: nCellsGroup2, isLoading: isLoadingGroup2 } =
-    useProcessedQueryGroupFilterDimensions(queryGroup2);
+  const {
+    isLoading,
+    setIsLoading,
+    queryGroup1,
+    queryGroup2,
+    canRunDifferentialExpression,
+    handleRunDifferentialExpression,
+    nCellsGroup1,
+    isLoadingGroup1,
+    nCellsGroup2,
+    isLoadingGroup2,
+  } = useConnect();
 
   return (
     <TwoPanelLayout>
